Cleanup and deal stages implementation
