+++ src/pages/opportunities/EditOpportunity.tsx
@@ -1,1028 +1,10 @@
-import React, { ChangeEvent, useEffect, useRef, useState } from 'react'
-import { useNavigate, useLocation } from 'react-router-dom'
-import {
-  TextField,
-  FormControl,
-  TextareaAutosize,
-  AccordionDetails,
-  Accordion,
-  AccordionSummary,
-  Typography,
-  Box,
-  MenuItem,
-  InputAdornment,
-  Chip,
-  Autocomplete,
-  FormHelperText,
-  IconButton,
-  Select,
-  Divider,
-  Button
-} from '@mui/material'
-import { useQuill } from 'react-quilljs'
-import 'quill/dist/quill.snow.css'
-import { OpportunityUrl } from '../../services/ApiUrls'
-import { fetchData } from '../../components/FetchData'
-import { CustomAppBar } from '../../components/CustomAppBar'
-import {
-  FaCheckCircle,
-  FaPlus,
-  FaTimes,
-  FaTimesCircle,
-  FaUpload
-} from 'react-icons/fa'
-import {
-  CustomPopupIcon,
-  RequiredSelect,
-  RequiredTextField
-} from '../../styles/CssStyled'
-import { FiChevronDown } from '@react-icons/all-files/fi/FiChevronDown'
-import { FiChevronUp } from '@react-icons/all-files/fi/FiChevronUp'
-import '../../styles/style.css'
-import { CURRENCY_CODES } from '../../utils/Constants'
-
-type FormErrors = {
-  name?: string[]
-  account?: string[]
-  amount?: string[]
-  currency?: string[]
-  stage?: string[]
-  teams?: string[]
-  lead_source?: string[]
-  probability?: string[]
-  description?: string[]
-  assigned_to?: string[]
-  contacts?: string[]
-  due_date?: string[]
-  tags?: string[]
-  opportunity_attachment?: string[]
-  file?: string[]
-  lead?: string[]
-}
-
-interface AssignedTo {
-  id: string
-  name: string
-}
-
-interface FormData {
-  name: string
-  account: string
-  amount: string
-  currency: string
-  stage: string
-  teams: string[]
-  lead_source: string
-  probability: number
-  description: string
-  assigned_to: AssignedTo[]
-  contacts: string[]
-  due_date: string
-  tags: string[]
-  opportunity_attachment: string | null
-  file: string | null
-  lead: string | null
-}
+import { useLocation } from 'react-router-dom'
+import { OpportunityForm } from './OpportunityForm'
+import { HTTP_METHODS } from '../../utils/Constants'
 
 export function EditOpportunity () {
-  const navigate = useNavigate()
-  const { state } = useLocation()
-  const { quill, quillRef } = useQuill()
-  const initialContentRef = useRef<string | null>(null)
-  const pageContainerRef = useRef<HTMLDivElement | null>(null)
-
-  const [hasInitialFocus, setHasInitialFocus] = useState(false)
-
-  const autocompleteRef = useRef<any>(null)
-  const [error, setError] = useState(false)
-  const [reset, setReset] = useState(false)
-  const [selectedContacts, setSelectedContacts] = useState<any[]>([])
-  const [selectedAssignTo, setSelectedAssignTo] = useState<any[]>([])
-  const [selectedTags, setSelectedTags] = useState<any[]>([])
-  const [selectedTeams, setSelectedTeams] = useState<any[]>([])
-  const [selectedCountry, setSelectedCountry] = useState<any[]>([])
-  const [leadSelectOpen, setLeadSelectOpen] = useState(false)
-  const [leadSourceSelectOpen, setLeadSourceSelectOpen] = useState(false)
-  const [statusSelectOpen, setStatusSelectOpen] = useState(false)
-  const [countrySelectOpen, setCountrySelectOpen] = useState(false)
-  const [currencySelectOpen, setCurrencySelectOpen] = useState(false)
-  const [contactSelectOpen, setContactSelectOpen] = useState(false)
-  const [accountSelectOpen, setAccountSelectOpen] = useState(false)
-  const [stageSelectOpen, setStageSelectOpen] = useState(false)
-  const [errors, setErrors] = useState<FormErrors>({})
-  const [formData, setFormData] = useState<FormData>({
-    name: '',
-    account: '',
-    amount: '',
-    currency: '',
-    stage: '',
-    teams: [],
-    lead_source: '',
-    probability: 1,
-    description: '',
-    assigned_to: [],
-    contacts: [],
-    due_date: '',
-    tags: [],
-    opportunity_attachment: null,
-    file: null,
-    lead: ''
-  })
-  const userOptions = state.users.map((k: any, i: any) => { return { id: k.id, label: k.user__email } })
-
-  useEffect(() => {
-    // Scroll to the top of the page when the component mounts
-    window.scrollTo(0, 0)
-    // Set focus to the page container after the Quill editor loads its content
-    if (quill && !hasInitialFocus) {
-      quill.on('editor-change', () => {
-        if (pageContainerRef.current) {
-          pageContainerRef.current.focus()
-          setHasInitialFocus(true) // Set the flag to true after the initial focus
-        }
-      })
-    }
-
-    // Cleanup: Remove event listener when the component unmounts
-    return () => {
-      if (quill) {
-        quill.off('editor-change')
-      }
-    }
-  }, [quill, hasInitialFocus])
-
-  useEffect(() => {
-    setFormData(state?.value)
-  }, [state?.id])
-
-  useEffect(() => {
-    if (reset) {
-      setFormData(state?.value)
-      if (quill && initialContentRef.current !== null) {
-        quill.clipboard.dangerouslyPasteHTML(initialContentRef.current)
-      }
-    }
-    if (state?.value?.assigned_to) {
-      let assignedToIds = state.value.assigned_to.map((k: any, i: any) => k.id)
-      setSelectedAssignTo(userOptions.filter((item: any) => assignedToIds.includes(item.id)))
-    }
-    return () => {
-      setReset(false)
-    }
-  }, [reset, quill, state?.value])
-
-  useEffect(() => {
-    if (quill && initialContentRef.current === null) {
-      // Save the initial state (HTML content) of the Quill editor only if not already saved
-      initialContentRef.current = formData.description
-      quill.clipboard.dangerouslyPasteHTML(formData.description)
-    }
-  }, [quill, formData.description])
-
-  const backbtnHandle = () => {
-    if (state?.edit) {
-      navigate('/app/opportunities')
-    } else {
-      navigate('/app/opportunities/opportunity-details', {
-        state: { opportunityId: state?.id, detail: true }
-      })
-    }
-  }
-  const handleChange2 = (title: any, val: any) => {
-    if (title === 'contacts') {
-      setFormData({
-        ...formData,
-        contacts: val.length > 0 ? val.map((item: any) => item.id) : []
-      })
-      setSelectedContacts(val)
-    } else if (title === 'assigned_to') {
-      setFormData({
-        ...formData,
-        assigned_to: val
-      })
-      setSelectedAssignTo(val)
-    } else if (title === 'tags') {
-      setFormData({
-        ...formData,
-        tags: val.length > 0 ? val.map((item: any) => item.id) : []
-      })
-      setSelectedTags(val)
-    } else if (title === 'teams') {
-      setFormData({
-        ...formData,
-        teams: val.length > 0 ? val.map((item: any) => item.id) : []
-      })
-      setSelectedTeams(val)
-    } else {
-      setFormData({ ...formData, [title]: val })
-    }
-  }
-  const handleChange = (e: any) => {
-    const { name, value, files, type, checked, id } = e.target
-    if (type === 'file') {
-      setFormData({ ...formData, [name]: e.target.files?.[0] || null })
-    } else if (type === 'checkbox') {
-      setFormData({ ...formData, [name]: checked })
-    } else {
-      let val = value
-      if (name === 'contacts') {
-        val = [value]
-      }
-      setFormData({ ...formData, [name]: val })
-    }
-  }
-  const handleFileChange = (event: any) => {
-    const file = event.target.files?.[0] || null
-    if (file) {
-      setFormData((prevData) => ({
-        ...prevData,
-        opportunity_attachment: file.name,
-        file: prevData.file
-      }))
-
-      const reader = new FileReader()
-      reader.onload = () => {
-        setFormData((prevData) => ({
-          ...prevData,
-          file: reader.result as string
-        }))
-      }
-      reader.readAsDataURL(file)
-    }
-  }
-  const emptyDescription = () => {
-    setFormData({ ...formData, description: '' })
-    if (quill) {
-      quill.clipboard.dangerouslyPasteHTML('')
-    }
-  }
-  const handleSubmit = (e: any) => {
-    e.preventDefault()
-    submitForm()
-  }
-  const submitForm = () => {
-    const Header = {
-      Accept: 'application/json',
-      'Content-Type': 'application/json',
-      Authorization: localStorage.getItem('Token'),
-      org: localStorage.getItem('org')
-    }
-
-    const data = {
-      name: formData.name,
-      account: formData.account,
-      amount: formData.amount,
-      currency: formData.currency,
-      stage: formData.stage,
-      teams: formData.teams,
-      lead_source: formData.lead_source,
-      probability: formData.probability,
-      description: formData.description,
-      assigned_to: formData.assigned_to.map((k: any, i: number) => k.id),
-      contacts: formData.contacts,
-      due_date: formData.due_date,
-      tags: formData.tags,
-      // opportunity_attachment: formData.opportunity_attachment,
-      opportunity_attachment: formData.file,
-      lead: formData.lead
-    }
-
-    fetchData(
-      `${OpportunityUrl}/${state?.id}/`,
-      'PUT',
-      JSON.stringify(data),
-      Header
+    const { state } = useLocation()
+    return (
+        <OpportunityForm state={state} httpReqMethod={HTTP_METHODS.PUT}/>
     )
-      .then((res: any) => {
-        // console.log('Form data:', res);
-        if (!res.error) {
-          resetForm()
-          navigate('/app/opportunities')
-        }
-        if (res.error) {
-          setError(true)
-          setErrors(res?.errors)
-        }
-      })
-      .catch(() => {})
-  }
-  const resetForm = () => {
-    setFormData({
-      name: '',
-      account: '',
-      amount: '',
-      currency: '',
-      stage: '',
-      teams: [],
-      lead_source: '',
-      probability: 1,
-      description: '',
-      assigned_to: [],
-      contacts: [],
-      due_date: '',
-      tags: [],
-      opportunity_attachment: null,
-      file: null,
-      lead: ''
-    })
-    setErrors({})
-    setSelectedContacts([])
-    setSelectedAssignTo([])
-    setSelectedTags([])
-    setSelectedTeams([])
-  }
-  const onCancel = () => {
-    // resetForm()
-    setReset(true)
-  }
-
-  const module = 'Opportunities'
-  const crntPage = 'Edit Opportunity'
-  const backBtn = state?.edit
-    ? 'Back To Opportunities'
-    : 'Back To OpportunityDetails'
-  return (
-    <Box sx={{ mt: '60px' }}>
-      <CustomAppBar
-        backbtnHandle={backbtnHandle}
-        module={module}
-        backBtn={backBtn}
-        crntPage={crntPage}
-        onCancel={onCancel}
-        onSubmit={handleSubmit}
-      />
-      <Box sx={{ mt: '120px' }}>
-        <form onSubmit={handleSubmit}>
-          <div style={{ padding: '10px' }}>
-            <div className="leadContainer">
-              <Accordion defaultExpanded style={{ width: '98%' }}>
-                <AccordionSummary
-                  expandIcon={<FiChevronDown style={{ fontSize: '25px' }} />}
-                >
-                  <Typography className="accordion-header">
-                    Opportunity Information
-                  </Typography>
-                </AccordionSummary>
-                <Divider className="divider" />
-                <AccordionDetails>
-                  <Box sx={{ width: '98%', color: '#1A3353', mb: 1 }}>
-                    <div className="fieldContainer">
-                      <div className="fieldSubContainer">
-                        <div className="fieldTitle">Name</div>
-                        <TextField
-                          ref={pageContainerRef}
-                          tabIndex={-1}
-                          name="account_name"
-                          value={formData.name}
-                          onChange={handleChange}
-                          style={{ width: '70%' }}
-                          size="small"
-                          helperText={errors?.name?.[0] ? errors?.name[0] : ''}
-                          error={!!errors?.name?.[0]}
-                        />
-                      </div>
-                      <div className="fieldSubContainer">
-                        <div className="fieldTitle">Amount</div>
-                        <TextField
-                          type={'number'}
-                          name="amount"
-                          value={formData.amount}
-                          onChange={handleChange}
-                          style={{ width: '70%' }}
-                          size="small"
-                          helperText={
-                            errors?.amount?.[0] ? errors?.amount[0] : ''
-                          }
-                          error={!!errors?.amount?.[0]}
-                        />
-                      </div>
-                    </div>
-                    <div className="fieldContainer2">
-                      <div className="fieldSubContainer">
-                        <div className="fieldTitle">Account</div>
-                        <FormControl sx={{ width: '70%' }}>
-                          <Select
-                            name="account"
-                            value={formData.account}
-                            open={accountSelectOpen}
-                            onClick={() =>
-                              setAccountSelectOpen(!accountSelectOpen)
-                            }
-                            IconComponent={() => (
-                              <div
-                                onClick={() =>
-                                  setAccountSelectOpen(!accountSelectOpen)
-                                }
-                                className="select-icon-background"
-                              >
-                                {accountSelectOpen ? (
-                                  <FiChevronUp className="select-icon" />
-                                ) : (
-                                  <FiChevronDown className="select-icon" />
-                                )}
-                              </div>
-                            )}
-                            className={'select'}
-                            onChange={handleChange}
-                            error={!!errors?.account?.[0]}
-                          >
-                            {state?.account?.length &&
-                              state?.account.map((option: any) => (
-                                <MenuItem key={option?.id} value={option?.id}>
-                                  {option?.name}
-                                </MenuItem>
-                              ))}
-                          </Select>
-                          <FormHelperText className="helperText">
-                            {errors?.currency?.[0] ? errors?.currency[0] : ''}
-                          </FormHelperText>
-                        </FormControl>
-                      </div>
-                      <div className="fieldSubContainer">
-                        <div className="fieldTitle">Currency</div>
-                        <FormControl sx={{ width: '70%' }}>
-                          <Select
-                            name="currency"
-                            value={formData.currency}
-                            open={currencySelectOpen}
-                            onClick={() =>
-                              setCurrencySelectOpen(!currencySelectOpen)
-                            }
-                            IconComponent={() => (
-                              <div
-                                onClick={() =>
-                                  setCurrencySelectOpen(!currencySelectOpen)
-                                }
-                                className="select-icon-background"
-                              >
-                                {currencySelectOpen ? (
-                                  <FiChevronUp className="select-icon" />
-                                ) : (
-                                  <FiChevronDown className="select-icon" />
-                                )}
-                              </div>
-                            )}
-                            className={'select'}
-                            onChange={handleChange}
-                            error={!!errors?.currency?.[0]}
-                          >
-                            {CURRENCY_CODES.map((option: any) => (
-                                <MenuItem key={option[0]} value={option[0]}>
-                                  {option[1]}
-                                </MenuItem>
-                              ))}
-                          </Select>
-                          <FormHelperText className="helperText">
-                            {errors?.currency?.[0] ? errors?.currency[0] : ''}
-                          </FormHelperText>
-                        </FormControl>
-                      </div>
-                    </div>
-                    <div className="fieldContainer2">
-                      <div className="fieldSubContainer">
-                        <div className="fieldTitle">Stage</div>
-                        <FormControl sx={{ width: '70%' }}>
-                          <RequiredSelect
-                            name="stage"
-                            value={formData.stage}
-                            open={stageSelectOpen}
-                            onClick={() => setStageSelectOpen(!stageSelectOpen)}
-                            IconComponent={() => (
-                              <div
-                                onClick={() =>
-                                  setStageSelectOpen(!stageSelectOpen)
-                                }
-                                className="select-icon-background"
-                              >
-                                {stageSelectOpen ? (
-                                  <FiChevronUp className="select-icon" />
-                                ) : (
-                                  <FiChevronDown className="select-icon" />
-                                )}
-                              </div>
-                            )}
-                            className={'select'}
-                            onChange={handleChange}
-                            error={!!errors?.stage?.[0]}
-                          >
-                            {state?.stage?.length &&
-                              state?.stage.map((option: any) => (
-                                <MenuItem key={option[0]} value={option[0]}>
-                                  {option[1]}
-                                </MenuItem>
-                              ))}
-                          </RequiredSelect>
-                          <FormHelperText className="helperText">
-                            {errors?.stage?.[0] ? errors?.stage[0] : ''}
-                          </FormHelperText>
-                        </FormControl>
-                      </div>
-                      <div className="fieldSubContainer">
-                        <div className="fieldTitle">Contact Name</div>
-                        <FormControl sx={{ width: '70%' }}>
-                          <RequiredSelect
-                            name="contacts"
-                            value={formData.contacts?.[0] || ''}
-                            open={contactSelectOpen}
-                            onClick={() =>
-                              setContactSelectOpen(!contactSelectOpen)
-                            }
-                            IconComponent={() => (
-                              <div
-                                onClick={() =>
-                                  setContactSelectOpen(!contactSelectOpen)
-                                }
-                                className="select-icon-background"
-                              >
-                                {contactSelectOpen ? (
-                                  <FiChevronUp className="select-icon" />
-                                ) : (
-                                  <FiChevronDown className="select-icon" />
-                                )}
-                              </div>
-                            )}
-                            className="select"
-                            onChange={handleChange}
-                            error={!!errors?.contacts?.[0]}
-                          >
-                            {state?.contacts?.length &&
-                              state?.contacts.map((option: any) => (
-                                <MenuItem
-                                  key={option?.id}
-                                  value={option?.id}
-                                >
-                                  {option?.first_name}
-                                </MenuItem>
-                              ))}
-                          </RequiredSelect>
-                          <FormHelperText className="helperText">
-                            {errors?.contacts?.[0]
-                              ? errors?.contacts[0]
-                              : ''}
-                          </FormHelperText>
-                        </FormControl>
-                      </div>
-                    </div>
-                    <div className="fieldContainer2">
-                      <div className="fieldSubContainer">
-                        <div className="fieldTitle">Lead Source</div>
-                        <FormControl sx={{ width: '70%' }}>
-                          <Select
-                            name="lead_source"
-                            value={formData.lead_source}
-                            open={leadSourceSelectOpen}
-                            onClick={() => setLeadSourceSelectOpen(!leadSourceSelectOpen)}
-                            IconComponent={() => (
-                              <div
-                                onClick={() =>
-                                  setLeadSourceSelectOpen(!leadSourceSelectOpen)
-                                }
-                                className="select-icon-background"
-                              >
-                                {leadSourceSelectOpen ? (
-                                  <FiChevronUp className="select-icon" />
-                                ) : (
-                                  <FiChevronDown className="select-icon" />
-                                )}
-                              </div>
-                            )}
-                            className={'select'}
-                            onChange={handleChange}
-                            error={!!errors?.lead_source?.[0]}
-                          >
-                            {state?.leadSource?.length &&
-                              state?.leadSource.map((option: any) => (
-                                <MenuItem key={option[0]} value={option[0]}>
-                                  {option[1]}
-                                </MenuItem>
-                              ))}
-                          </Select>
-                          <FormHelperText className="helperText">
-                            {errors?.lead_source?.[0]
-                              ? errors?.lead_source[0]
-                              : ''}
-                          </FormHelperText>
-                        </FormControl>
-                      </div>
-                      <div className="fieldSubContainer">
-                        <div className="fieldTitle">Probability</div>
-                        <TextField
-                          type={'number'}
-                          name="probability"
-                          value={formData.probability}
-                          onChange={handleChange}
-                          style={{ width: '70%' }}
-                          size="small"
-                          helperText={
-                            errors?.probability?.[0]
-                              ? errors?.probability[0]
-                              : ''
-                          }
-                          error={!!errors?.probability?.[0]}
-                        />
-                      </div>
-                    </div>
-                    <div className="fieldContainer2">
-                      <div className="fieldSubContainer">
-                        <div className="fieldTitle">Assign To</div>
-                        <FormControl
-                          error={!!errors?.assigned_to?.[0]}
-                          sx={{ width: '70%' }}
-                        >
-                          <Autocomplete
-                            multiple
-                            value={selectedAssignTo}
-                            limitTags={2}
-                            options={userOptions || []}
-                            getOptionLabel={(option: any) =>
-                              state.users ? option?.label : option
-                            }
-                            onChange={(e: any, value: any) =>
-                              handleChange2('assigned_to', value)
-                            }
-                            size="small"
-                            filterSelectedOptions
-                            renderTags={(value, getTagProps) =>
-                              value.map((option, index) => (
-                                <Chip
-                                  deleteIcon={
-                                    <FaTimes style={{ width: '9px' }} />
-                                  }
-                                  sx={{
-                                    backgroundColor: 'rgba(0, 0, 0, 0.08)',
-                                    height: '18px'
-                                  }}
-                                  variant="outlined"
-                                  label={
-                                    state.users ? option?.label : option
-                                  }
-                                  {...getTagProps({ index })}
-                                />
-                              ))
-                            }
-                            popupIcon={
-                              <CustomPopupIcon>
-                                <FaPlus className="input-plus-icon" />
-                              </CustomPopupIcon>
-                            }
-                            renderInput={(params) => (
-                              <TextField
-                                {...params}
-                                placeholder="Add Users"
-                                InputProps={{
-                                  ...params.InputProps,
-                                  sx: {
-                                    '& .MuiAutocomplete-popupIndicator': {
-                                      '&:hover': { backgroundColor: 'white' }
-                                    },
-                                    '& .MuiAutocomplete-endAdornment': {
-                                      mt: '-8px',
-                                      mr: '-8px'
-                                    }
-                                  }
-                                }}
-                              />
-                            )}
-                          />
-                          <FormHelperText>
-                            {errors?.assigned_to?.[0] || ''}
-                          </FormHelperText>
-                        </FormControl>
-                      </div>
-                      <div className="fieldSubContainer">
-                        <div className="fieldTitle">Due Date</div>
-                        <TextField
-                          type={'date'}
-                          name="due_date"
-                          value={formData.due_date}
-                          onChange={handleChange}
-                          style={{ width: '70%' }}
-                          size="small"
-                          helperText={
-                            errors?.due_date?.[0] ? errors?.due_date[0] : ''
-                          }
-                          error={!!errors?.due_date?.[0]}
-                          sx={{
-                            '& input[type="date"]::-webkit-calendar-picker-indicator':
-                              {
-                                backgroundColor: 'whitesmoke',
-                                padding: '13px',
-                                marginRight: '-15px'
-                              }
-                          }}
-                        />
-                      </div>
-                    </div>
-                    <div className="fieldContainer2">
-                      <div className="fieldSubContainer">
-                        <div className="fieldTitle">Tags</div>
-                        <FormControl
-                          error={!!errors?.tags?.[0]}
-                          sx={{ width: '70%' }}
-                        >
-                          <Autocomplete
-                            // ref={autocompleteRef}
-                            value={selectedTags}
-                            multiple
-                            limitTags={5}
-                            options={state.tags || []}
-                            // options={state.contacts ? state.contacts.map((option: any) => option) : ['']}
-                            getOptionLabel={(option: any) => option}
-                            onChange={(e: any, value: any) =>
-                              handleChange2('tags', value)
-                            }
-                            size="small"
-                            filterSelectedOptions
-                            renderTags={(value, getTagProps) =>
-                              value.map((option, index) => (
-                                <Chip
-                                  deleteIcon={
-                                    <FaTimes style={{ width: '9px' }} />
-                                  }
-                                  sx={{
-                                    backgroundColor: 'rgba(0, 0, 0, 0.08)',
-                                    height: '18px'
-                                  }}
-                                  variant="outlined"
-                                  label={option}
-                                  {...getTagProps({ index })}
-                                />
-                              ))
-                            }
-                            popupIcon={
-                              <CustomPopupIcon>
-                                <FaPlus className="input-plus-icon" />
-                              </CustomPopupIcon>
-                            }
-                            renderInput={(params) => (
-                              <TextField
-                                {...params}
-                                placeholder="Add Tags"
-                                InputProps={{
-                                  ...params.InputProps,
-                                  sx: {
-                                    '& .MuiAutocomplete-popupIndicator': {
-                                      '&:hover': { backgroundColor: 'white' }
-                                    },
-                                    '& .MuiAutocomplete-endAdornment': {
-                                      mt: '-8px',
-                                      mr: '-8px'
-                                    }
-                                  }
-                                }}
-                              />
-                            )}
-                          />
-                          <FormHelperText>
-                            {errors?.tags?.[0] || ''}
-                          </FormHelperText>
-                        </FormControl>
-                      </div>
-                      <div className="fieldSubContainer">
-                        <div className="fieldTitle">Lead Attachment</div>
-                        <TextField
-                          name="opportunity_attachment"
-                          value={formData.opportunity_attachment}
-                          InputProps={{
-                            endAdornment: (
-                              <InputAdornment position="end">
-                                <IconButton
-                                  disableFocusRipple
-                                  disableTouchRipple
-                                  sx={{
-                                    width: '40px',
-                                    height: '40px',
-                                    backgroundColor: 'whitesmoke',
-                                    borderRadius: '0px',
-                                    mr: '-13px',
-                                    cursor: 'pointer'
-                                  }}
-                                >
-                                  <label htmlFor="icon-button-file">
-                                    <input
-                                      hidden
-                                      accept="image/*"
-                                      id="icon-button-file"
-                                      type="file"
-                                      name="opportunity_attachment"
-                                      onChange={(e: any) => {
-                                        //  handleChange(e);
-                                        handleFileChange(e)
-                                      }}
-                                    />
-                                    <FaUpload
-                                      color="primary"
-                                      style={{
-                                        fontSize: '15px',
-                                        cursor: 'pointer'
-                                      }}
-                                    />
-                                  </label>
-                                </IconButton>
-                              </InputAdornment>
-                            )
-                          }}
-                          sx={{ width: '70%' }}
-                          size="small"
-                          helperText={
-                            errors?.opportunity_attachment?.[0]
-                              ? errors?.opportunity_attachment[0]
-                              : ''
-                          }
-                          error={!!errors?.opportunity_attachment?.[0]}
-                        />
-                      </div>
-                    </div>
-                    <div className="fieldContainer2">
-                      <div className="fieldSubContainer">
-                        <div className="fieldTitle" style={{ width: '35%' }}>
-                          Teams
-                        </div>
-                        <FormControl
-                          error={!!errors?.teams?.[0]}
-                          sx={{ width: '85%' }}
-                        >
-                          <Autocomplete
-                            // ref={autocompleteRef}
-                            value={selectedTeams}
-                            multiple
-                            limitTags={5}
-                            options={state.teams || []}
-                            // options={state.contacts ? state.contacts.map((option: any) => option) : ['']}
-                            getOptionLabel={(option: any) => option}
-                            onChange={(e: any, value: any) =>
-                              handleChange2('teams', value)
-                            }
-                            size="small"
-                            filterSelectedOptions
-                            renderTags={(value, getTagProps) =>
-                              value.map((option, index) => (
-                                <Chip
-                                  deleteIcon={
-                                    <FaTimes style={{ width: '9px' }} />
-                                  }
-                                  sx={{
-                                    backgroundColor: 'rgba(0, 0, 0, 0.08)',
-                                    height: '18px'
-                                  }}
-                                  variant="outlined"
-                                  label={option}
-                                  {...getTagProps({ index })}
-                                />
-                              ))
-                            }
-                            popupIcon={
-                              <CustomPopupIcon>
-                                <FaPlus className="input-plus-icon" />
-                              </CustomPopupIcon>
-                            }
-                            renderInput={(params) => (
-                              <TextField
-                                {...params}
-                                placeholder="Add Teams"
-                                InputProps={{
-                                  ...params.InputProps,
-                                  sx: {
-                                    '& .MuiAutocomplete-popupIndicator': {
-                                      '&:hover': { backgroundColor: 'white' }
-                                    },
-                                    '& .MuiAutocomplete-endAdornment': {
-                                      mt: '-8px',
-                                      mr: '-8px'
-                                    }
-                                  }
-                                }}
-                              />
-                            )}
-                          />
-                          <FormHelperText>
-                            {errors?.teams?.[0] || ''}
-                          </FormHelperText>
-                        </FormControl>
-                      </div>
-                      <div className="fieldSubContainer">
-                        <div className="fieldTitle">Lead</div>
-                        <FormControl sx={{ width: '70%' }}>
-                          <Select
-                            name="lead"
-                            value={formData.lead}
-                            open={leadSelectOpen}
-                            onClick={() => setLeadSelectOpen(!leadSelectOpen)}
-                            IconComponent={() => (
-                              <div
-                                onClick={() =>
-                                  setLeadSelectOpen(!leadSelectOpen)
-                                }
-                                className="select-icon-background"
-                              >
-                                {leadSelectOpen ? (
-                                  <FiChevronUp className="select-icon" />
-                                ) : (
-                                  <FiChevronDown className="select-icon" />
-                                )}
-                              </div>
-                            )}
-                            className={'select'}
-                            onChange={handleChange}
-                            error={!!errors?.lead?.[0]}
-                          >
-                            {state?.leads?.length &&
-                              state?.leads.map((option: any) => (
-                                <MenuItem key={option?.id} value={option?.id}>
-                                  {option?.title}
-                                </MenuItem>
-                              ))}
-                          </Select>
-                          <FormHelperText className="helperText">
-                            {errors?.lead?.[0] || ''}
-                          </FormHelperText>
-                        </FormControl>
-                      </div>
-                    </div>
-                  </Box>
-                </AccordionDetails>
-              </Accordion>
-            </div>
-            {/* Description details  */}
-            <div className="leadContainer">
-              <Accordion defaultExpanded style={{ width: '98%' }}>
-                <AccordionSummary
-                  expandIcon={<FiChevronDown style={{ fontSize: '25px' }} />}
-                >
-                  <Typography className="accordion-header">
-                    Description
-                  </Typography>
-                </AccordionSummary>
-                <Divider className="divider" />
-                <AccordionDetails>
-                  <Box
-                    sx={{ width: '100%', mb: 1 }}
-                    component="form"
-                    noValidate
-                    autoComplete="off"
-                  >
-                    <div className="DescriptionDetail">
-                      <div className="descriptionTitle">Description</div>
-                      <div style={{ width: '100%', marginBottom: '3%' }}>
-                        <div ref={quillRef} />
-                      </div>
-                    </div>
-                    <Box
-                      sx={{
-                        display: 'flex',
-                        flexDirection: 'row',
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                        mt: 1.5
-                      }}
-                    >
-                      <Button
-                        className="header-button"
-                        onClick={onCancel}
-                        size="small"
-                        variant="contained"
-                        startIcon={
-                          <FaTimesCircle
-                            style={{
-                              fill: 'white',
-                              width: '16px',
-                              marginLeft: '2px'
-                            }}
-                          />
-                        }
-                        sx={{
-                          backgroundColor: '#2b5075',
-                          ':hover': { backgroundColor: '#1e3750' }
-                        }}
-                      >
-                        Cancel
-                      </Button>
-                      <Button
-                        className="header-button"
-                        onClick={handleSubmit}
-                        variant="contained"
-                        size="small"
-                        startIcon={
-                          <FaCheckCircle
-                            style={{
-                              fill: 'white',
-                              width: '16px',
-                              marginLeft: '2px'
-                            }}
-                          />
-                        }
-                        sx={{ ml: 1 }}
-                      >
-                        Save
-                      </Button>
-                    </Box>
-                  </Box>
-                </AccordionDetails>
-              </Accordion>
-            </div>
-          </div>
-        </form>
-      </Box>
-    </Box>
-  )
 }
