+++ src/pages/opportunities/OpportunityDetails.tsx
@@ -19,6 +19,7 @@ import { FaPlus, FaStar } from 'react-icons/fa'
 import FormateTime from '../../components/FormateTime'
 import { Label } from '../../components/Label'
 import { COUNTRIES } from '../../utils/Constants'
+import { Stage, StageConnector, OpportunityStages } from '../../components/OpportunityStages'
 
 export const formatDate = (dateString: any) => {
   const options: Intl.DateTimeFormatOptions = {
@@ -155,7 +156,6 @@ export const OpportunityDetails = (props: any) => {
     }
     fetchData(`${OpportunityUrl}/${id}/`, 'GET', null as any, Header)
       .then((res) => {
-        console.log(res, 'edd')
         if (!res.error) {
           setOpportunityDetails(res?.opportunity_obj)
           setUsers(res?.users)
@@ -163,7 +163,6 @@ export const OpportunityDetails = (props: any) => {
         }
       })
       .catch((err) => {
-        // console.error('Error:', err)
         ;<Snackbar
           open={err}
           autoHideDuration={4000}
@@ -241,12 +240,21 @@ export const OpportunityDetails = (props: any) => {
   const module = 'Opportunities'
   const crntPage = 'Opportunity Details'
   const backBtn = 'Back To Opportunities'
-  console.log(state, 'oppdetail')
   const userDetail = (userId: any, e: React.MouseEvent<HTMLAnchorElement>) => {
     e.preventDefault()
     navigate('/app/users/user-details', { state: { userId, detail: true } })
   }
 
+  const leadStages = []
+  for (const stage of state?.stage) {
+    let stateInfo: {title: string, color?: string } = { title: stage[0] }
+    if (stage[0] === 'CLOSED WON') {
+      stateInfo.color = 'rgb(8, 86, 8)'
+    } else if (stage[0] === 'CLOSED LOST') {
+      stateInfo.color = 'rgb(132, 26, 26)'
+    }
+    leadStages.push(stateInfo)
+  }
   return (
     <Box sx={{ mt: '60px' }}>
       <div>
@@ -262,11 +270,12 @@ export const OpportunityDetails = (props: any) => {
             mt: '110px',
             p: '20px',
             display: 'flex',
-            flexDirection: 'row',
+            flexDirection: 'column',
             justifyContent: 'space-between'
           }}
         >
-          <Box sx={{ width: '65%' }}>
+          {/* <Box sx={{ width: '65%' }}> */}
+          <Box sx={{ width: '100%' }}>
             <Box
               sx={{
                 borderRadius: '10px',
@@ -326,6 +335,8 @@ export const OpportunityDetails = (props: any) => {
                   </div>
                 </div>
               </div>
+              <OpportunityStages orderedStageList={ leadStages } currentStage={opportunityDetails?.stage}/>
+              
               <div
                 style={{
                   padding: '20px',
@@ -347,7 +358,7 @@ export const OpportunityDetails = (props: any) => {
                 >
                   {opportunityDetails?.tags?.length
                     ? opportunityDetails?.tags.map((tagData: any) => (
-                        <Label tags={tagData} />
+                        <Label tags={tagData.name} />
                       ))
                     : ''}
                 </Stack>
@@ -462,27 +473,32 @@ export const OpportunityDetails = (props: any) => {
                   <div className="title2">Assigned Users</div>
                   <div className="title3" style={{
                       display: 'flex',
-                      flexDirection: 'row',
+                      flexDirection: 'column',
                       justifyContent: 'flex-start',
-                      alignItems: 'center',
-                      marginRight: '15px'
+                      alignItems: 'flex-start'
+                      // marginRight: '15px'
                     }}>
-                {opportunityDetails?.assigned_to?.length
-                  ? opportunityDetails.assigned_to.map(
-                      (item: any, i: any) => (
-                              <Avatar
-                                key={i}
-                                src={item?.user_details?.profile_pic}
-                                alt={item?.user_details?.email}
-                              />
-                      )
-                    )
-                  : ('----')}
-                {opportunityDetails?.assigned_to?.length
+                    {opportunityDetails?.assigned_to?.length
+                      ? opportunityDetails.assigned_to.map(
+                          (item: any, i: any) => (
+                              <div style={{ display: 'flex', alignItems: 'center', gap: '15px' }}>
+                                <Avatar
+                                  key={i}
+                                  sx={{ width: 24, height: 24 }}
+                                  src={item?.user_details?.profile_pic}
+                                  alt={item?.user_details?.email}
+                                />
+                                <p>{item?.user_details?.email}</p>
+                              </div>
+                                  
+                          )
+                        )
+                      : ('----')}
+                {/* {opportunityDetails?.assigned_to?.length
                   ? opportunityDetails.assigned_to.map(
                       (item: any, i: any) => (item?.user_details?.email)
                     ).join(',')
-                  : ''}
+                  : ''} */}
                   </div>
                 </div>
                 <div style={{ width: '32%' }}>
@@ -574,7 +590,8 @@ export const OpportunityDetails = (props: any) => {
               </div>
             </Box>
           </Box>
-          <Box sx={{ width: '34%' }}>
+          {/* ATTACHMENTS SECTION BEGIN */}
+          <Box sx={{ width: '100%' }}>
             <Box
               sx={{
                 borderRadius: '10px',
@@ -639,6 +656,7 @@ export const OpportunityDetails = (props: any) => {
               </div>
             </Box>
           </Box>
+          {/* ATTACHMENTS SECTION END */}
         </Box>
       </div>
     </Box>
